requirements.txt
# Django
Django==4.2.7
djangorestframework==3.14.0

# Database
psycopg2-binary==2.9.7

# SMS
africastalking==3.0.0

# Authentication
mozilla-django-oidc==3.0.0

# Production
gunicorn==21.2.0
whitenoise==6.5.0

# Testing
pytest==7.4.2
pytest-django==4.7.0
pytest-cov==4.1.0
coverage==7.3.2

# Environment
python-dotenv==1.0.0


requests==2.31.0

# Date/time
python-dateutil==2.8.2